Fix error in not considering the result of the second argument of <*
Previous implementation was discarding the Result[T] for <*, which meant that writing
```nim
many(digit) <* eof
```
did also match `8394   483` because the eof would fail but such result would be discarded.

Instead what we want is to also try to parse the second result and if it fails then we return a failure,
otherwise we return only the first correctly parsed result.